+++ play/src/main/scala/com/sandinh/couchbase/PlayCBCluster.scala
@@ -1,20 +1,23 @@
 package com.sandinh.couchbase
 
+import java.lang
 import javax.inject.{Inject, Singleton}
-import play.api.Configuration
+
 import play.api.inject.ApplicationLifecycle
+
 import scala.concurrent.duration._
 import com.sandinh.rx.Implicits._
-import scala.concurrent.ExecutionContext.Implicits.global
+import com.typesafe.config.Config
+
+import scala.concurrent.Future
 
 @Singleton
-class PlayCBCluster @Inject() (cfg: Configuration, lifecycle: ApplicationLifecycle) extends CBCluster(cfg.underlying) {
+class PlayCBCluster @Inject() (cfg: Config, lifecycle: ApplicationLifecycle) extends CBCluster(cfg) {
   /** convention val for using with play.api.inject.ApplicationLifecycle#addStopHook */
-  val disconnectFuture = () =>
+  val disconnectFuture: () => Future[lang.Boolean] = () =>
     asJava.disconnect()
       .timeout(env.disconnectTimeout, MILLISECONDS)
       .toFuture
-      .map(_ => ())
 
   lifecycle.addStopHook(disconnectFuture)
 }